requirements.txt
kalshi-python>=2.1
python-dotenv>=1.0
fastapi>=0.115
uvicorn>=0.30
nba_api>=1.11
boto3>=1.34
pandas>=2.0
